Move dive loop TEDs to their own file
