+++ src/openrct2/libopenrct2.vcxproj
@@ -558,6 +558,7 @@
     <ClInclude Include="ride\rtd\water\SplashBoats.h" />
     <ClInclude Include="ride\rtd\water\SubmarineRide.h" />
     <ClInclude Include="ride\ted\PitchAndRoll.h" />
+    <ClInclude Include="ride\ted\TED.DiveLoop.h" />
     <ClInclude Include="ride\ted\TED.SBend.h" />
     <ClInclude Include="ride\ted\TED.ZeroG.h" />
     <ClInclude Include="ride\ted\TrackElementDescriptor.h" />
